Added button to show milestone report from node grading view. #2299
