+++ src/main/webapp/wise5/classroomMonitor/classroomMonitorComponents/milestones/milestoneDetails/milestoneDetails.ts
@@ -136,6 +136,7 @@ class MilestoneDetailsController {
 const MilestoneDetails = {
   bindings: {
     milestone: '<',
+    hideStudentWork: '<',
     onShowWorkgroup: '&',
     onVisitNodeGrading: '&'
   },
@@ -185,29 +186,35 @@ const MilestoneDetails = {
           <compile data="$ctrl.milestone.generatedReport"></compile>
         </div>
       </section>
-      <section ng-if="$ctrl.milestone.generatedRecommendations && $ctrl.milestone.isReportAvailable"
-          class="md-whiteframe-1dp gray-lightest-bg">
-        <md-tabs md-dynamic-height>
-          <md-tab label="{{ ::'recommendations' | translate }}" md-on-select="$ctrl.saveTabSelectedEvent('MilestoneRecommendationTabSelected')">
-            <div class="milestone-details__section">
-              <compile data="$ctrl.milestone.generatedRecommendations"></compile>
-            </div>
-          </md-tab>
-          <md-tab label="{{ ::'studentWork' | translate }}" md-on-select="$ctrl.saveTabSelectedEvent('MilestoneStudentWorkTabSelected')">
-            <div class="milestone-details__section">
-              <node-grading-view node-id="$ctrl.milestone.nodeId"
-                                 milestone="$ctrl.milestone"></node-grading-view>
-            </div>
-          </md-tab>
-        </md-tabs>
+      <section ng-if="$ctrl.hideStudentWork && $ctrl.milestone.generatedRecommendations && $ctrl.milestone.isReportAvailable"
+               class="milestone-details__section md-whiteframe-1dp">
+        <div class="milestone-details__header primary md-body-2 gray-lightest-bg">{{ ::'recommendations' | translate }}</div>
+        <compile data="$ctrl.milestone.generatedRecommendations"></compile>
       </section>
-      <section ng-if="!$ctrl.milestone.generatedRecommendations && $ctrl.milestone.isReportAvailable"
-          class="milestone-details__section md-whiteframe-1dp">
-        <div class="milestone-details__header primary md-body-2 gray-lightest-bg">{{ ::'studentWork' | translate }}</div>
-        <node-grading-view node-id="$ctrl.milestone.nodeId"
-                           milestone="$ctrl.milestone"></node-grading-view>
+      <section ng-if="!$ctrl.hideStudentWork" class="md-whiteframe-1dp gray-lightest-bg">
+        <div ng-if="$ctrl.milestone.generatedRecommendations && $ctrl.milestone.isReportAvailable">
+          <md-tabs md-dynamic-height>
+            <md-tab label="{{ ::'recommendations' | translate }}" md-on-select="$ctrl.saveTabSelectedEvent('MilestoneRecommendationTabSelected')">
+              <div class="milestone-details__section">
+                <compile data="$ctrl.milestone.generatedRecommendations"></compile>
+              </div>
+            </md-tab>
+            <md-tab label="{{ ::'studentWork' | translate }}" md-on-select="$ctrl.saveTabSelectedEvent('MilestoneStudentWorkTabSelected')">
+              <div class="milestone-details__section">
+                <node-grading-view node-id="$ctrl.milestone.nodeId"
+                                   milestone="$ctrl.milestone"></node-grading-view>
+              </div>
+            </md-tab>
+          </md-tabs>
+        </div>
+        <div ng-if="!$ctrl.milestone.generatedRecommendations && $ctrl.milestone.isReportAvailable"
+                 class="milestone-details__section md-whiteframe-1dp">
+          <div class="milestone-details__header primary md-body-2 gray-lightest-bg">{{ ::'studentWork' | translate }}</div>
+          <node-grading-view node-id="$ctrl.milestone.nodeId"
+                             milestone="$ctrl.milestone"></node-grading-view>
+        </div>
       </section>
-      <section ng-if="!$ctrl.milestone.generatedRecommendations && !$ctrl.milestone.isReportAvailable"
+      <section ng-if="!$ctrl.milestone.isReportAvailable"
           class="milestone-details__section md-whiteframe-1dp">
         <div class="milestone-details__header primary md-body-2 gray-lightest-bg">{{ ::'studentCompletion' | translate }}</div>
         <ng-include src="'completion'"></ng-include>
